requirements.txt
openai
python-telegram-bot==13.15